+++ mapie/regression.py
@@ -456,6 +456,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         agg_function = self._check_agg_function(self.agg_function)
         X, y = indexable(X, y)
         y = _check_y(y)
+        sample_weight = cast(Optional[NDArray], sample_weight)
         self.n_features_in_ = check_n_features_in(X, cv, estimator)
         sample_weight, X, y = check_null_weight(sample_weight, X, y)
         self.conformity_score_function_ = check_conformity_score(
@@ -654,7 +655,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         y_pred_low = np.column_stack(
             [
                 np_nanquantile(
-                    lower_bounds,
+                    lower_bounds.astype(float),
                     _alpha,
                     axis=1,
                     method="lower",
@@ -665,7 +666,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         y_pred_up = np.column_stack(
             [
                 np_nanquantile(
-                    upper_bounds,
+                    upper_bounds.astype(float),
                     1 - _alpha,
                     axis=1,
                     method="higher",